packages/data-ingestor/pyproject.toml
[project]
name = "data-ingestor"
version = "0.1.0"
description = "Python project for processing and ingesting data into the MIRPR-PPI RAG System"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "convex>=0.7.0",
    "docling>=2.58.0",
    "python-dotenv>=1.1.1",
]

[dependency-groups]
dev = [
    "pytest>=8.4.2",
    "ruff>=0.14.1",
]
